Update sidebar to match project image
Used a mix of grid and flexbox as I found flexbox was easier to center
the svg icons with the text. Also, decided to use spans for svg icons
as it seemed easiest to place it directly in the html with a span.

I got in the zone and forgot to commit earlier so a few changes for the
sidebar are lumped along with changes for the header.