Create main section cards
Used a mix of grid and flexbox again as I find the flexbox easier to
just align one thing in a box or one-dimensionally. I initially used a
box shadow to mimic the gold bar on the left of the cards, but it was
better to just use border-left, especially since I needed to use the box
shadow for the shadow under the card.